Extract publishing_api_draft_exporter
